server/pyproject.toml
[project]
name = "flood-resilience-backend"
version = "1.0.0"
description = "Flood Resilience System Backend for Sri Lanka"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "geoalchemy2>=0.14.0",
    "pgvector>=0.2.4",
    
    # Redis & Celery
    "redis>=5.0.0",
    "celery>=5.3.0",
    
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    
    # HTTP & Utils
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "orjson>=3.9.0",
    
    # Storage
    "boto3>=1.34.0",
    "aiofiles>=23.2.0",
    
    # Telemetry
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.43b0",
    
    # Validation
    "email-validator>=2.1.0",
    "phonenumbers>=8.13.0",
    
    # Geo
    "shapely>=2.0.0",
    "geopy>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
